階層整理、セッション認証の追加
