+++ back/src/api/app/todo/todolist.go
@@ -1,41 +1,12 @@
 package todo
 
 import (
-	"time"
-
 	"../../domain"
 	"../../infra"
 	"../../infra/table"
-	"github.com/lib/pq"
 )
 
-func ToPatchGoal(id int) (err error) {
-	db, err := infra.DBConnect()
-	if err != nil {
-		return err
-	}
-
-	var todo table.TodoList
-
-	err = db.Table("todo_lists").
-		Where("id = ?", id).
-		First(&todo).
-		Error
-
-	if err != nil {
-		return err
-	}
-
-	todo.IsGoaled = true
-	todo.GoaledAt = pq.NullTime{Time: time.Now(), Valid: true}
-
-	db.Save(&todo)
-
-	return
-
-}
-
-func ToGetAllGoal(limit int, page int, order string) (out []allTodoArray, err error) {
+func ToGetAll(limit int, page int, order string) (out []allTodoArray, err error) {
 
 	db, err := infra.DBConnect()
 
@@ -47,13 +18,13 @@ func ToGetAllGoal(limit int, page int, order string) (out []allTodoArray, err er
 
 	base := db.Table("todo_lists").
 		Select("todo_lists.id, todo_lists.Content, todo_lists.user_id, todo_lists.created_at, todo_lists.last_achieved, todo_lists.is_deleted, todo_lists.is_goaled, users.name, users.handle_name, users.img").
-		Where("todo_lists.is_deleted = ? and todo_lists.is_goaled = ?", false, true).
+		Where("todo_lists.is_deleted = ? and todo_lists.is_goaled = ?", false, false).
 		Joins("left join users on users.ID = todo_lists.user_id").
 		Limit(limit).
 		Offset(limit * (page - 1))
 
 	err = base.
-		Order("todo_lists.last_achieved").
+		Order("todo_lists.last_achieved desc").
 		Scan(&rows).
 		Error
 
@@ -96,7 +67,7 @@ func ToGetAllGoal(limit int, page int, order string) (out []allTodoArray, err er
 
 }
 
-func ToGetOneGoal(name string, order string) (out userTodoArray, err error) {
+func ToGetOneUser(name string, order string) (out userTodoArray, err error) {
 	db, err := infra.DBConnect()
 
 	if err != nil {
@@ -133,10 +104,10 @@ func ToGetOneGoal(name string, order string) (out userTodoArray, err error) {
 
 	base := db.Table("todo_lists").
 		Select("id, user_id, content, created_at, last_achieved, is_deleted, is_goaled").
-		Where("user_id = ? and is_deleted = ? and is_goaled = ?", userID, false, true)
+		Where("user_id = ? and is_deleted = ? and is_goaled = ?", userID, false, false)
 
 	err = base.
-		Order("last_achieved").
+		Order("last_achieved desc").
 		Scan(&rows).
 		Error
